backend/requirements.txt
# Core Framework
fastapi[all]==0.115.12
uvicorn[standard]==0.34.2
websockets<15.0
python-dotenv==1.1.0
pydantic-settings==2.9.1

# Telephony
twilio==9.6.1

# Speech Processing
faster-whisper==1.1.1
# piper-tts==1.2.0
numpy==2.2.5
soundfile==0.13.1

# Intelligence - Local Models
ollama==0.5.1

# Intelligence - Cloud Models
groq==0.25.0

# Intelligence - Web Search
duckduckgo-search==7.5.3

# Memory & Database
supabase==2.15.2

# Emotion Detection
textblob==0.19.0

# Utilities
python-multipart==0.0.20
httpx==0.28.1
loguru==0.7.3
